Cleanup pass 5: function-size NOLINTs + real-bug fixes + counter logging
clang-tidy: 168 → 124 (-44). Cumulative from bootstrap: 3171 → 124
(96.1% reduction).

Three classes of fix:

1. readability-function-size: 27 NOLINTs (inline at decl line)

   Each calculator's main accumulator function (BiotSavart, McConnell,
   Mopac variants, Larsen, Tripeptide, dihedral, FullSystemReader,
   Trajectory, OrcaRunLoader, etc.) NOLINT'd with rationale per
   CLAUDE.md "no helpers for helpers' sake". The function-size limit
   is a heuristic; thesis-physics code where one function = one
   per-atom-per-bond accumulation is harder to read split.

2. Real-bug fixes — unused/dead variables (12 sites)

   unused-variable (6):
     SolventEnvironment.h::Dipole mid_H — computed but unused (μ
       calculation doesn't reference it). Removed; comment preserved.
     WaterFieldResult.cpp d_O — sqrt(d_O_sq) but only d_O_sq used.
       Removed.
     AIMNet2Result.cpp WriteFeatures protein — declared, never used.
       Removed.
     CoulombResult.cpp::SamplePointE protein — declared, never used.
       Removed.
     DsspResult.cpp SS_CLASSES "HGIEBTS C" string — mnemonic for the
       column-order comment; never accessed. Removed; comment kept
       (preserves the documentation, drops the dead const).
     HaighMallionResult.cpp::WriteFeatures written counter — declared
       at function entry but never incremented; function returns 3
       literally. Removed.

   unused-but-set (6) — pattern was "counter accumulating diagnostic
   info that wasn't logged":
     HBondResult filtered_out → added to OperationLog::Info at the
       end (same pattern as McConnellResult in pass-2,
       MutationDeltaResult in pass-3).
     MopacMcConnellResult filtered_out → added to OperationLog::Info.
     EeqResult n_fallback → added to OperationLog::Info as
       fallback_params= (S uses real, Unknown uses fallback).
     WaterFieldResult n_beyond_cutoff → dead counter (cutoff-filter
       short-circuits without using count for anything). Removed.
     FullSystemReader global_offset → set but unused; the per-moltype
       walk doesn't need a global index. Removed.
     CalculatorConfig line_num → parser counter never used in error
       messages. Removed (errors don't include line numbers; a
       future cleanup could thread it through if needed).

3. Real-bug fixes — implicit widening (5 sites)

   int*int multiplication in flat-index calculations assigned to
   size_t. Cast left-operand to std::size_t:
     PiQuadrupoleResult.cpp:282    per_type_T2[i * 40 + t * 5 + c]
     DispersionResult.cpp:417      per_type_T2[i * 40 + t * 5 + c]
     MopacMcConnellResult.cpp:324  cat_T2[i * 25 + c * 5 + m]
     HaighMallionResult.cpp:430    per_type_T2[i * 40 + t * 5 + c]
     DsspResult.cpp:294            base = i * 12 + k * 3
   Same overflow-catch class as the bulk-fix in pass-1
   (LarsenHBondGrid + BiotSavartResult).

Tests: 158/158 in focused regression (WaterField + HBond + McConnell +
Mopac + Eeq + Dssp + HaighMallion + FullSystemReader + Coulomb +
AIMNet2 + FullPipeline + PiQuad + Dispersion families). Build green.

Top remaining categories (124 findings):
  10  bugprone-branch-clone               case-by-case
   9  performance-inefficient-string-concatenation  perf
   8  concurrency-mt-unsafe               wrapper-NOLINT
   6  misc-const-correctness              auto-fix straggler
   6  misc-unused-parameters              review
   6  bugprone-implicit-widening (remaining)
   5  cert-flp30-c                        loop induction FP
   5  bugprone-unchecked-optional-access  false positives (refactor)
   5  bugprone-argument-comment           review
   5  cert-env33-c                        NOLINT
   ... long tail

+++ src/PiQuadrupoleResult.cpp
@@ -279,7 +279,7 @@ int PiQuadrupoleResult::WriteFeatures(const ProteinConformation& conf,
         for (int t = 0; t < 8; ++t) {
             per_type_T0[i*8 + t] = ca.per_type_pq_scalar_sum[t];
             for (int c = 0; c < 5; ++c) {
-                per_type_T2[i*40 + t*5 + c] = ca.per_type_pq_T2_sum[t][c];
+                per_type_T2[i * 40 + static_cast<std::size_t>(t) * 5 + c] = ca.per_type_pq_T2_sum[t][c];
 }
         }
     }
